Update the CI routine.

+++ build.nim
@@ -1,4 +1,4 @@
-import os, strutils, json
+import os, parseopt, strutils, json
 import flavors / [slim, regular, onbuild]
 
 proc isDefault(props: JsonNode): bool = props.getOrDefault("default").getBVal
@@ -107,26 +107,35 @@ when isMainModule:
     bases = config["bases"]
     versions = config["versions"]
 
+  var targets: seq[string] = @[]
+
+  for kind, key, val in getopt():
+    case kind
+    of cmdArgument: targets.add key
+    else: discard
+
   for version in versions.pairs:
-    for base in bases.pairs:
-      for flavor in flavors:
-        let tags = getTags(version, base, flavor)
+    if len(targets) == 0 or version.key in targets:
+      for base in bases.pairs:
+        for flavor in flavors:
+          let tags = getTags(version, base, flavor)
 
-        echo "Building $#... " % tags[0]
-        generateDockerfile(version.key, base.key, flavor, labels)
-        buildImage(tags, tagPrefix)
-        removeFile("Dockerfile")
-        echo "Done!"
-
-        # Anything before this is broken and too old to fix.
-        if version.key >= "0.16.0":
-          echo "Testing $#... " % tags[0]
-          testImage(
-            "$#:$#" % [tagPrefix, tags[0]],
-            flavor
-          )
+          echo "Building $#... " % tags[0]
+          generateDockerfile(version.key, base.key, flavor, labels)
+          buildImage(tags, tagPrefix)
+          removeFile("Dockerfile")
           echo "Done!"
 
-        echo "Pushing $#..." % tags[0]
-        pushImage(tags, tagPrefix)
-        echo "Done!"
+          # Anything before this is broken and too old to fix.
+          if version.key >= "0.16.0":
+            echo "Testing $#... " % tags[0]
+            testImage(
+              "$#:$#" % [tagPrefix, tags[0]],
+              flavor
+            )
+            echo "Done!"
+
+
+          echo "Pushing $#..." % tags[0]
+          pushImage(tags, tagPrefix)
+          echo "Done!"
